Make AuthGuard and RolesGuard global

+++ advisor-backend/src/assessment/assessment.controller.ts
@@ -7,7 +7,6 @@ import {
   Param,
   Delete,
   ParseIntPipe,
-  UseGuards,
   ForbiddenException,
 } from '@nestjs/common';
 import {
@@ -24,8 +23,6 @@ import { UpdateAssessmentDto } from './dto/update-assessment.dto';
 import { AssessmentDto } from './dto/assessment.dto';
 import { FeedbackDto } from './dto/feedback.dto';
 import { SaveCheckpointDto } from './dto/save-checkpoint.dto';
-import { RolesGuard } from '../common/guards/roles.guard';
-import { AuthGuard } from '@nestjs/passport';
 import { Roles } from '../common/decorators/roles.decorator';
 import { Role, User } from '@prisma/client';
 import AuthUser from '../common/decorators/auth-user.decorator';
@@ -44,7 +41,7 @@ export class AssessmentController {
    * @param createAssessmentDto Assessment information
    * @returns created assessment
    */
-  @Post()
+  @Post('')
   @ApiConflictResponse({
     description: 'Assessment with this name and type already exists',
   })
@@ -68,6 +65,7 @@ export class AssessmentController {
     type: AssessmentDto,
     isArray: true,
   })
+  @Roles(Role.ADMIN)
   findAll() {
     return this.assessmentService.findAll();
   }
@@ -78,7 +76,6 @@ export class AssessmentController {
    * @returns AssessmentDto[] List of all assessments
    */
   @Get('my')
-  @UseGuards(AuthGuard('jwt'))
   @ApiResponse({
     description: 'Found assessment of USER',
     type: AssessmentDto,
@@ -96,7 +93,16 @@ export class AssessmentController {
   @Get(':assessment_id')
   @ApiResponse({ description: 'Assessment', type: AssessmentDto })
   @ApiNotFoundResponse({ description: 'Assessment not found' })
-  findOne(@Param('assessment_id', ParseIntPipe) id: number) {
+  async findOne(
+    @Param('assessment_id', ParseIntPipe) id: number,
+    @AuthUser() user: User
+  ) {
+    const assessment = await this.assessmentService.userInAssessment(id, user);
+
+    if (!assessment) {
+      throw new ForbiddenException();
+    }
+
     return this.assessmentService.findOne(id);
   }
 
@@ -112,10 +118,17 @@ export class AssessmentController {
   @ApiConflictResponse({
     description: 'Assessment with this name and type already exists',
   })
-  update(
+  async update(
     @Param('assessment_id', ParseIntPipe) id: number,
-    @Body() updateAssessmentDto: UpdateAssessmentDto
+    @Body() updateAssessmentDto: UpdateAssessmentDto,
+    @AuthUser() user: User
   ) {
+    const assessment = await this.assessmentService.userInAssessment(id, user);
+
+    if (!assessment) {
+      throw new ForbiddenException();
+    }
+
     return this.assessmentService.update(id, updateAssessmentDto);
   }
 
@@ -127,7 +140,16 @@ export class AssessmentController {
   @Delete(':assessment_id')
   @ApiResponse({ description: 'Assessment', type: AssessmentDto })
   @ApiNotFoundResponse({ description: 'Assessment not found' })
-  delete(@Param('assessment_id', ParseIntPipe) id: number) {
+  async delete(
+    @Param('assessment_id', ParseIntPipe) id: number,
+    @AuthUser() user: User
+  ) {
+    const assessment = await this.assessmentService.userInAssessment(id, user);
+
+    if (!assessment) {
+      throw new ForbiddenException();
+    }
+
     return this.assessmentService.delete(id);
   }
 
@@ -139,7 +161,16 @@ export class AssessmentController {
   @Post(':assessment_id/complete')
   @ApiResponse({ description: 'Assessment', type: AssessmentDto })
   @ApiNotFoundResponse({ description: 'Assessment not found' })
-  complete(@Param('assessment_id', ParseIntPipe) id: number) {
+  async complete(
+    @Param('assessment_id', ParseIntPipe) id: number,
+    @AuthUser() user: User
+  ) {
+    const assessment = await this.assessmentService.userInAssessment(id, user);
+
+    if (!assessment) {
+      throw new ForbiddenException();
+    }
+
     return this.assessmentService.complete(id);
   }
 
@@ -150,7 +181,6 @@ export class AssessmentController {
    * @returns Checkpoint saved
    */
   @Post(':assessment_id/save')
-  @UseGuards(AuthGuard('jwt'))
   @ApiOkResponse({ description: 'Checkpoint saved' })
   @ApiNotFoundResponse({ description: 'Assessment not found' })
   async saveCheckpointAnswer(
@@ -177,7 +207,6 @@ export class AssessmentController {
    * @returns Saved checkpoints
    */
   @Get(':assessment_id/save')
-  @UseGuards(AuthGuard('jwt'))
   @ApiResponse({
     description: 'Saved checkpoints',
     type: SaveCheckpointDto,
@@ -206,7 +235,6 @@ export class AssessmentController {
    * @returns updated Assessment object
    */
   @Post(':assessment_id/feedback')
-  @UseGuards(AuthGuard('jwt'), RolesGuard)
   @Roles(Role.ASSESSOR)
   @ApiResponse({ description: 'Assessment', type: AssessmentDto })
   @ApiNotFoundResponse({ description: 'Assessment not found' })
